+++ test/seq/runtests.jl
@@ -3288,45 +3288,44 @@ end
 
     @testset "2bit" begin
         buffer = IOBuffer()
-        writer = TwoBitWriter(buffer, ["chr1", "chr2"])
+        writer = TwoBit.Writer(buffer, ["chr1", "chr2"])
         chr1 = dna"ACGTNN"
         chr2 = dna"N"^100 * dna"ACGT"^100 * dna"N"^100
         write(writer, SeqRecord("chr1", chr1))
         write(writer, SeqRecord("chr2", chr2))
         seekstart(buffer)
-        reader = TwoBitReader(buffer)
+        reader = TwoBit.Reader(buffer)
         @test length(reader) == 2
-        @test reader["chr1"].name == "chr1"
-        @test reader["chr1"].seq == chr1
-        @test reader["chr2"].name == "chr2"
-        @test reader["chr2"].seq == chr2
-        @test reader["chr1"].name == "chr1"
-        @test reader["chr1"].seq == chr1
+        @test TwoBit.seqnames(reader) == ["chr1", "chr2"]
+        @test TwoBit.sequence(reader["chr1"]) == chr1
+        @test TwoBit.sequence(reader["chr2"]) == chr2
+        @test TwoBit.sequence(reader["chr1"]) == chr1
         @test_throws KeyError reader["chr10"]
-        @test reader[1].name == "chr1"
-        @test reader[2].name == "chr2"
         @test_throws BoundsError reader[3]
 
         function check_2bit_parse(filename)
-            stream = open(TwoBitReader, filename)
-            @test eltype(stream) == SeqRecord{ReferenceSequence,Vector{UnitRange{Int}}}
+            stream = open(TwoBit.Reader, filename)
+            @test eltype(stream) === TwoBit.Record
             # read from a stream
-            for record in stream end
+            for record in stream
+                @test hassequence(record) == TwoBit.hassequence(record) == true
+                @test TwoBit.sequence(ReferenceSequence, record) == TwoBit.sequence(DNASequence, record)
+            end
             close(stream)
 
             # round trip
             buffer = IOBuffer()
-            reader = open(TwoBitReader, filename)
-            writer = TwoBitWriter(buffer, reader.names)
-            expected_entries = Any[]
-            for record in reader
-                write(writer, record)
+            reader = open(TwoBit.Reader, filename)
+            writer = TwoBit.Writer(buffer, TwoBit.seqnames(reader))
+            expected_entries = TwoBit.Record[]
+            for (name, record) in zip(TwoBit.seqnames(reader), reader)
+                write(writer, SeqRecord(name, TwoBit.sequence(record), TwoBit.maskedblocks(record)))
                 push!(expected_entries, record)
             end
 
-            read_entries = Any[]
+            read_entries = TwoBit.Record[]
             seekstart(buffer)
-            for record in TwoBitReader(buffer)
+            for record in TwoBit.Reader(buffer)
                 push!(read_entries, record)
             end