+++ NEWS.md
@@ -12,6 +12,7 @@ Bio.jl v0.5.0 Release Notes
 * Computing genomic distances using the MASH algorithm is introduced ([#415]).
 * IO APIs are reorganized into submodules:
     * `Bio.Seq.FASTAReader` => `Bio.Seq.FASTA.Reader` ([#413]).
+    * `Bio.Seq.TwoBitReader` => `Bio.Seq.TwoBit.Reader` ([#431]).
     * `Bio.Intervals.BEDReader` => `Bio.Intervals.BED.Reader` ([#418]).
 * Overloaded `Base.intersect` methods for intervals are removed. Use the `eachoverlap` function exported from `Bio.Intervals` instead ([#426]).
 * A reader for the ABIF format by Applied Biosystems is introduced ([#353]).
@@ -27,3 +28,4 @@ Bio.jl v0.5.0 Release Notes
 [#415]: https://github.com/BioJulia/Bio.jl/pull/415
 [#418]: https://github.com/BioJulia/Bio.jl/pull/418
 [#426]: https://github.com/BioJulia/Bio.jl/pull/426
+[#431]: https://github.com/BioJulia/Bio.jl/pull/431